backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pandas==2.2.2
groq==0.8.0
alpha_vantage==2.3.1
pytrends==4.9.2
python-multipart==0.0.9
requests==2.32.3
python-dotenv==1.0.1
